+++ microagent/tools/amqp.py
@@ -148,7 +148,8 @@ class AMQPBroker(AbstractQueueBroker):
             :param \*\*kwargs: dict, other basic_publish options
         '''  # noqa: W605
         async with ChannelContext(self) as channel:
-            await channel.basic_publish(message, routing_key=name, exchange_name=exchange, **kwargs)
+            await channel.basic_publish(message.encode(), routing_key=name,
+                exchange_name=exchange, **kwargs)
 
     def _on_amqp_error(self, name: str, exception: Exception):
         self.log.warning('Catch AMPQ exception %s on queue "%s"', exception, name)
@@ -211,7 +212,7 @@ class AMQPBroker(AbstractQueueBroker):
 
     def _amqp_wrapper(self, consumer: Consumer) -> Callable:
         async def _wrapper(channel, body, envelope, properties):
-            data = consumer.queue.deserialize(body)
+            data = self.prepared_data(consumer, body)
 
             if not data:
                 self.log.debug('Calling %s by %s without data', consumer, consumer.queue.name)